feat(initial-design): add album card and reorganize components
